+++ mesh-exam/domain/api/src/main/java/io/flaterlab/meshexam/domain/create/model/ExamModel.kt
@@ -4,5 +4,5 @@ data class ExamModel(
     val id: String,
     val name: String,
     val type: String,
-    val durationInMin: Long,
+    val durationInMin: Int,
 )
