+++ mesh-exam/feature/discover/src/main/java/io/flaterlab/meshexam/presentation/discover/dvo/AvailableExamDvo.kt
@@ -6,5 +6,5 @@ data class AvailableExamDvo(
     override val id: String,
     override val name: String,
     val hostName: String,
-    override val durationInMin: Long,
+    override val durationInMin: Int,
 ) : ExamListAdapter.ExamItem
